Move MTOSHipmentUpdater service object to own file and test file
